re-wrote the mine check
Check for mines now creates a list when reading in the CSV at the start of the script. If keyword 'mine' is in the layer nickname, the associated path is stored in a list. The list is called with the mine function. A user could now add as many mine layers as desired without having to alter the py file.